Fix canvas tainted error - revert CORS change that broke downloads
The previous 'fix' in commit f24ed71 actually broke downloads by NOT setting crossOrigin for Google Drive images. This caused the canvas to be tainted and unable to export.

Reverted to the original working code that sets crossOrigin='anonymous' for ALL images, which prevents canvas tainting when used with Cross-Origin-Embedder-Policy: 'unsafe-none'